gave the code a good scrub. wouldn't call it dirty, but it's not full of sand anymore

+++ src/glbrick/LineSpec.java
@@ -56,7 +56,7 @@ public class LineSpec implements BrickSpec {
 
 	
 	public DrawnObject toDrawnObject() {
-		float[] temp = new float[]{1f,1f,1f}; 
+		double[] temp = new double[]{1f,1f,1f}; 
 		return new DrawnObject(vertices, temp);
 	}
